Draw path + show game info
- Game
  - Show game info
    - Used when the webscoket connection is not established
  - Draw path when matching tiles
    - To do that, we need to change the way that we render the board:
use the `gameState` to store the updatedBoard and `board`` to render what we want.
Reason to do that?
When doing a match, we send a match request to server. Then the server
send us two messages: 1.updated game state and 2. the valid path for
linking the tiles.
What we want is the matched tiles and the path disappear together,
so we need do make it "disappear" after having the path from server, i.e.,
updating the board and the path at the same time.
The approach before we were updating the board when we received the
updated game state.